# modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        Node:
            description:
                https://www.schema.org/Node
            properties:
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                id:
                    format: int64
                    type: integer
                initial:
                    type: boolean
                label:
                    readOnly: true
                    type: string
                name:
                    type: string
                processId:
                    format: int64
                    type: integer
                processVersion:
                    type: string
                terminal:
                    type: boolean
                type:
                    type: string
            type: object
...
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
...
    "/processes/{processId}/nodes":
        get:
            operationId: getProcessNodesPage
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Node"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Node"
                                type: array
                    description:
                        ""
            tags: ["Node"]
        post:
            operationId: postProcessNode
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Node"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Node"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Node"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Node"
                    description:
                        ""
            tags: ["Node"]
...
    "/processes/{processId}/{processVersion}/nodes/{nodeId}":
        delete:
            operationId: deleteProcessNode
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: processVersion
                  required: true
                  schema:
                      type: string
                - in: path
                  name: nodeId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["Node"]
...